+++ android/app/src/main/kotlin/com/famereani/tutor_time/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.ajent
+package com.famereani.tutor_time
 
 import io.flutter.embedding.android.FlutterActivity
